+++ java-client-sdk/lib/src/main/java/industries/dot/brease/SDKConfiguration.java
@@ -14,7 +14,7 @@ class SDKConfiguration {
 	public int serverIdx = 0;
 	public String language = "java";
 	public String openapiDocVersion = "0.1.0";
-	public String sdkVersion = "0.3.0";
-	public String genVersion = "2.37.0";
+	public String sdkVersion = "0.2.3";
+	public String genVersion = "2.37.2";
 	
 }